chore: 파일 분리 및 URL 원본으로
