+++ Waving-iOS/Data/Network/URLString.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 struct Server {
-    //static let baseURL = "http://118.67.130.127:8080" //우리 서버
-    static let baseURL = "https://reqres.in/api" // API 테스트용
+    static let baseURL = "http://118.67.130.127:8080" //우리 서버
+    //static let baseURL = "https://reqres.in/api" // API 테스트용
 }
